+++ .vscode/launch.json
@@ -5,11 +5,11 @@
     "version": "0.2.0",
     "configurations": [
 
+
         {
             "name": "Python: Current File",
             "type": "python",
             "request": "launch",
-            // "program": "${file}",
             "program": "${workspaceFolder}/run.py",
             "console": "integratedTerminal",
             "justMyCode": true
