+++ run.py
@@ -1,6 +1,11 @@
 from datetime import date
 
-from app.src.arConverter import convert
+from arConverter.src.arConverter import convert
+
+def main():
+    targetDate = date.today()
+    arDate = convert(targetDate)
+    print(arDate.weekdayDate())
 
 if __name__ == "__main__":
     targetDate = date.today()
